Add operator_key into config
It will be announced along with the other oasis info if included.

There is also a tremndous amount of cleanup here separating concerns
into reasonable places.

Also changed some defaults in the config for my own reasons.

Use of the operator key to manage the social graph will be coming soon.